requirements.txt
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
pathlib>=1.0.1